requirements.txt
fastapi==0.103.1
opencv-python==4.8.0.76
python-dotenv==1.0.0
requests==2.31.0
ultralytics==8.0.168
uvicorn==0.23.2
